HashedExpression

Documentation

exteriorDerivative

Arguments

:: DimensionType d

=> Set String
Variables

-> Expression d R
Expression

-> Expression d Covector

Exterior derivative

derivativeAllVars :: DimensionType d => Expression d R -> Expression d Covector

Take derivative with all vars

data R

Type representation of elements in the 1D, 2D, 3D, ... grid

data C

data Covector

data Expression d et

Expression with 2 phantom types (dimension and num type)

data Scalar

Type representation of vector dimension

class PowerOp a b | a -> b where

Classes for operations so that both Expression and Pattern (in HashedPattern) can implement

Methods

(^) :: a -> b -> a infixl 8

class PiecewiseOp a b where

Methods

piecewise :: [Double] -> a -> [b] -> b

class VectorSpaceOp a b where

Minimal complete definition

scale

Methods

scale :: a -> b -> b infixl 8

(*.) :: a -> b -> b infixl 8

class FTOp a b | a -> b where

Methods

ft :: a -> b

class ComplexRealOp r c | r -> c, c -> r where

Methods

(+:) :: r -> r -> c infixl 6

xRe :: c -> r

xIm :: c -> r

class RotateOp k a | a -> k where

Methods

rotate :: k -> a -> a

class InnerProductSpaceOp a b c | a b -> c where

Methods

(<.>) :: a -> b -> c infixl 8

constant :: Double -> Expression Scalar R

constant1D :: forall n. KnownNat n => Double -> Expression n R

constant2D :: forall m n. (KnownNat m, KnownNat n) => Double -> Expression '(m, n) R

constant3D :: forall m n p. (KnownNat m, KnownNat n, KnownNat p) => Double -> Expression '(m, n, p) R

variable :: String -> Expression Scalar R

Create primitive expressions

variable1D :: forall n. KnownNat n => String -> Expression n R

Create primitive expressions using Nat kind

variable2D :: forall m n. (KnownNat m, KnownNat n) => String -> Expression '(m, n) R

variable3D :: forall m n p. (KnownNat m, KnownNat n, KnownNat p) => String -> Expression '(m, n, p) R

sum :: (Foldable t, Num a) => t a -> a

The sum function computes the sum of the numbers of a structure.

product :: (Foldable t, Num a) => t a -> a

The product function computes the product of the numbers of a structure.

normalize :: (DimensionType d, ElementType et) => Expression d et -> Expression d et

Normalize an expression

class Evaluable d rc output | d rc -> output where

These should be commented properly.

Methods

eval :: ValMaps -> Expression d rc -> output