HashedExpression.Internal.Expression

Documentation

data R

Type representation of elements in the 1D, 2D, 3D, ... grid

data C

data Covector

data ET

Data representation of element type

Constructors

R

C

Covector

data Node

Node type

Constructors

Var String

DVar String

Const Double

Sum ET Args

Mul ET Args

Power Int Arg

Neg ET Arg

Scale ET Arg Arg

Div Arg Arg

Sqrt Arg

Sin Arg

Cos Arg

Tan Arg

Exp Arg

Log Arg

Sinh Arg

Cosh Arg

Tanh Arg

Asin Arg

Acos Arg

Atan Arg

Asinh Arg

Acosh Arg

Atanh Arg

RealImag Arg Arg

RealPart Arg

ImagPart Arg

InnerProd ET Arg Arg

Piecewise [Double] ConditionArg [BranchArg]

Rotate RotateAmount Arg

ReFT Arg

ImFT Arg

TwiceReFT Arg

TwiceImFT Arg

type Internal = (Shape, Node)

Internal Shape: Shape of the expression we can reconstruct the type of the Expression

type ExpressionMap = IntMap Internal

Hash map of all subexpressions

data Expression d et

Expression with 2 phantom types (dimension and num type)

Constructors

Expression

Fields

exIndex :: Int

exMap :: ExpressionMap

data Scalar

Type representation of vector dimension

class Dimension d

class Dimension d => ToShape d where

Methods

toShape :: Proxy d -> Shape

type DimensionType d = (Dimension d, ToShape d)

class ElementType et

Classes as constraints

class ElementType et => NumType et

class VectorSpace d et s

class VectorSpace d s s => InnerProductSpace d s

class PowerOp a b | a -> b where

Classes for operations so that both Expression and Pattern (in HashedPattern) can implement

Methods

(^) :: a -> b -> a infixl 8

class PiecewiseOp a b where

Methods

piecewise :: [Double] -> a -> [b] -> b

class VectorSpaceOp a b where

Minimal complete definition

scale

Methods

scale :: a -> b -> b infixl 8

(*.) :: a -> b -> b infixl 8

class FTOp a b | a -> b where

Methods

ft :: a -> b

class ComplexRealOp r c | r -> c, c -> r where

Methods

(+:) :: r -> r -> c infixl 6

xRe :: c -> r

xIm :: c -> r

class RotateOp k a | a -> k where

Methods

rotate :: k -> a -> a

type Shape = [Int]

Shape type: [] --> scalar [n] --> 1D with size n [n, m] --> 2D with size n × m [n, m, p] --> 3D with size n × m × p

type RotateAmount = [Int]

Rotation in each dimension. | Property: the length of this must match the dimension of the data

type Arg = Int

type Args = [Int]

Args - list of indices of arguments in the ExpressionMap

type BranchArg = Int

type ConditionArg = Int

class InnerProductSpaceOp a b c | a b -> c where

Methods

(<.>) :: a -> b -> c infixl 8