HashedExpression.Internal.Inner

Description

Inner Hashed functions, there are functions as how to combine expression with options

Documentation

data D_

Placeholder for any dimension type

data ET_

Placeholder for any element type

unwrap :: Expression d et -> (ExpressionMap, Int)

wrap :: (ExpressionMap, Int) -> Expression d et

apply :: OperationOption -> [(ExpressionMap, Int)] -> (ExpressionMap, Int)

The apply function that is used everywhere

mulMany :: [(ExpressionMap, Int)] -> (ExpressionMap, Int)

General multiplication and sum

sumMany :: [(ExpressionMap, Int)] -> (ExpressionMap, Int)

hasShape :: OperationOption -> Shape -> OperationOption

applyBinary :: HasCallStack => OperationOption -> Expression d1 et1 -> Expression d2 et2 -> Expression d3 et3

applyNary :: HasCallStack => OperationOption -> [Expression d1 et1] -> Expression d2 et2

applyConditionAry :: OperationOption -> Expression d et1 -> [Expression d et2] -> Expression d et2

binary :: (Arg -> Arg -> Node) -> OperationOption

binary operations

binaryET :: (ET -> Arg -> Arg -> Node) -> ElementOutcome -> OperationOption

conditionAry :: (ConditionArg -> [BranchArg] -> Node) -> OperationOption

branch operation

applyUnary :: HasCallStack => OperationOption -> Expression d1 et1 -> Expression d2 et2

unary :: (Arg -> Node) -> OperationOption

unary operations

unaryET :: (ET -> Arg -> Node) -> ElementOutcome -> OperationOption

naryET :: (ET -> Args -> Node) -> ElementOutcome -> OperationOption

data ElementOutcome

Constructors

ElementSpecific ET

ElementDefault

const_ :: Shape -> Double -> Change

num_ :: Double -> Change

product_ :: [Change] -> Change

diffConst :: Shape -> Double -> ExpressionDiff

topologicalSort :: (ExpressionMap, Int) -> [Int]

Topological sort the expression map, all the dependencies will appear before the depended node, and all unreachable nodes will be ignored

topologicalSortManyRoots :: (ExpressionMap, [Int]) -> [Int]

Topological sort, but with many roots

data ExpressionDiff

Modification will return an ExpressionDiff instead of the whole Expression to speed things up What ExpressionDiff stands for?

Constructors

ExpressionDiff

Fields

extraEntries :: ExpressionMap

newRootId :: Int

type Transformation = (ExpressionMap, Int) -> (ExpressionMap, Int)

Transformation type, we can combine them, chain them, apply them n times using nest, ...

type Change = ExpressionMap -> ExpressionDiff

Change w.r.t original expression map

type Modification = (ExpressionMap, Int) -> Change

Modification type, given an expression, it will give a difference (i.e, extraEntries in the ExpressionMap, and the new index of the root expression) between the modified and original expression

fromModification :: Modification -> (ExpressionMap, Int) -> ExpressionDiff

withContext :: ExpressionMap -> Change -> ExpressionDiff

just :: Int -> Change

sum_ :: [Change] -> Change

toTransformation :: ((ExpressionMap, Int) -> ExpressionDiff) -> Transformation

multipleTimes :: Int -> Transformation -> Transformation

Apply maximum k times, or stop if the expression doesn't change

data OperandOrder

Operand order in the operation

Constructors

Reorder

NoReorder

toRecursive :: OperandOrder -> ((ExpressionMap, Int) -> ExpressionDiff) -> (ExpressionMap, Int) -> ExpressionDiff

Turn a a recursive one, i.e, apply rules to every node in the expression bottom up

combineChildrenDiffs :: OperandOrder -> ExpressionMap -> Int -> [ExpressionDiff] -> ExpressionDiff

Same node type (Mul, Sum, Negate, ...), but children may changed, now make the same node type with new children and return the combined difference

removeUnreachable :: Transformation

Remove unreachable nodes

mulManyDiff :: ExpressionMap -> [ExpressionDiff] -> ExpressionDiff

sumManyDiff :: ExpressionMap -> [ExpressionDiff] -> ExpressionDiff

applyDiff :: ExpressionMap -> OperationOption -> [ExpressionDiff] -> ExpressionDiff

noChange :: Int -> ExpressionDiff

The ExpressionDiff corresponding to no change in this node

expressionVarNodes :: (DimensionType d, ElementType et) => Expression d et -> [(String, Int)]

All variables in the Expression

varNodesWithId :: ExpressionMap -> [(String, Int)]

containsFTNode :: ExpressionMap -> Bool

If expression map contains any FT node

introduceZeroPartialDerivatives :: [(String, Shape)] -> Expression Scalar Covector -> Expression Scalar Covector

A function might be treated as function of variables that not appears in the function itself (like constraint of optimization problem, so we want to pad zero partial derivative) (( ... ) dx + ( .... ) dy) [z, t] ---> (( ... ) dx + ( ... ) dy + 0dz + 0dt)