HashedExpression.Internal.Pattern

Description

(c) 2014 Christopher Kumar Anand Helper functions/instances to make pattern guards involving Expressions easier to read.

Documentation

type Capture = Int

Pattern for normalizier

type ListCapture = Int

type PowerCapture = Int

type RotateAmountCapture = Int

data PatternList

List holes to captures many elements

Constructors

PListHole (Pattern -> Pattern) ListCapture

data PatternPower

Pattern power to capture the alpha in x ^ alpha

Constructors

PPowerHole PowerCapture

PPowerConst Int

PPowerMul PatternPower PatternPower

PPowerSum PatternPower PatternPower

data PatternRotateAmount

PatternRotateAmount to capture the amount in (rotate amount x)

Constructors

PRotateAmountHole RotateAmountCapture

PRotateAmountSum PatternRotateAmount PatternRotateAmount

PRotateAmountNegate PatternRotateAmount

data Pattern

Constructors

PHole Capture

PRef Int

PHead PatternList

PSumList PatternList

PMulList PatternList

PConst Double

PScalarConst Double

PSum [Pattern]

PMul [Pattern]

PNeg Pattern

PScale Pattern Pattern

PDiv Pattern Pattern

PSqrt Pattern

PSin Pattern

PCos Pattern

PTan Pattern

PExp Pattern

PLog Pattern

PSinh Pattern

PCosh Pattern

PTanh Pattern

PAsin Pattern

PAcos Pattern

PAtan Pattern

PAsinh Pattern

PAcosh Pattern

PAtanh Pattern

PRealImag Pattern Pattern

PRealPart Pattern

PImagPart Pattern

PInnerProd Pattern Pattern

PPiecewise Pattern PatternList

PMulRest ListCapture [Pattern]

PSumRest ListCapture [Pattern]

PPower Pattern PatternPower

PRotate PatternRotateAmount Pattern

PReFT Pattern

PImFT Pattern

PTwiceReFT Pattern

PTwiceImFT Pattern

class MulRestOp a b c | a b -> c where

Methods

(~*) :: a -> b -> c infixl 7

class SumRestOp a b c | a b -> c where

Methods

(~+) :: a -> b -> c infixl 6

mapL :: (Pattern -> Pattern) -> PatternList -> PatternList

Pattern List

reFT :: Pattern -> Pattern

Discrete fourier transform

imFT :: Pattern -> Pattern

twiceReFT :: Pattern -> Pattern

twiceImFT :: Pattern -> Pattern

data GuardedPattern

Guarded patterns for normalizier

Constructors

GP Pattern Condition

type Substitution = (GuardedPattern, Pattern)

fromSubstitution :: Substitution -> (ExpressionMap, Int) -> ExpressionDiff

Turn HashedPattern to a normalizier

(|.~~~~~~>) :: Pattern -> Pattern -> Substitution infix 0

Helper to make pattern and replacement without condition

(~~~~~~>) :: GuardedPattern -> Pattern -> Substitution infix 0

(|.) :: Pattern -> Condition -> GuardedPattern infixl 1

type Condition = (ExpressionMap, Int) -> Match -> Bool

(&&.) :: Condition -> Condition -> Condition infixl 8

isNot :: Condition -> Condition

(||.) :: Condition -> Condition -> Condition

isScalar :: Pattern -> Condition

isConst :: Pattern -> Condition

isNotConst :: Pattern -> Condition

isReal :: Pattern -> Condition

isComplex :: Pattern -> Condition

isCovector :: Pattern -> Condition

sameElementType :: [Pattern] -> Condition

allTheSame :: PatternList -> Condition

isDVar :: Pattern -> Condition

zeroAmount :: PatternRotateAmount -> Condition

sameAmount :: PatternRotateAmount -> PatternRotateAmount -> Condition

p :: Pattern

one :: Pattern

zero :: Pattern

powerOne :: PatternPower

powerZero :: PatternPower

alpha :: PatternPower

scalarOne :: Pattern

scalarZero :: Pattern

scalar :: Double -> Pattern

num :: Double -> Pattern

ys :: PatternList

xs :: PatternList

sum :: PatternList -> Pattern

product :: PatternList -> Pattern

piecewise_ :: Pattern -> PatternList -> Pattern

branches :: PatternList

headL :: PatternList -> Pattern

restOfProduct :: Pattern

restOfSum :: Pattern

amount :: PatternRotateAmount

matchList :: ExpressionMap -> [Int] -> PatternList -> Maybe Match

Matches all nodes in the expression to see if they all match the PatternList, if they match, return the inner actual nodes e.g: matchList [a + (b * x), a + (b * y), a + (b * z)] (PatternList: (a + (b * _)) ---> Just [x, y, z] matchList [x, y, z, t] (PatternList: (_)) = Just [x, y, z, t] matchList [1 + x, 2 + x, y + x] (PatternList: (a + _)) = Nothing (not the same for all)

type NodeId = Int

Match is the result we get when match an Expression against a Pattern

type PowerValue = Int

data Match

Constructors

Match

Fields

capturesMap :: Map Capture NodeId

listCapturesMap :: Map ListCapture [NodeId]

powerCapturesMap :: Map PowerCapture PowerValue

rotateAmountCapturesMap :: Map RotateAmountCapture RotateAmount

emptyMatch :: Match

unionMatch :: Match -> Match -> Match

match :: (ExpressionMap, Int) -> Pattern -> Maybe Match

Match an expression with a pattern, return the map between capture hole to the actual node e.g: match (Expression: (a(3243) + b(32521)) (PatternNormal:(x(1) + y(2)) --> ({1 -> 3243, 2 -> 32521}, {}) match (Expression sum(a(3243), b(32521), c(21321)) (PatternNormal:(sum(each(1))) --> ({}, {1 -> [3243, 32521, 21321]})

turnToPattern :: (Pattern -> Pattern) -> Int -> Pattern

buildFromPatternPower :: Match -> PatternPower -> PowerValue

buildFromPatternRotateAmount :: Match -> PatternRotateAmount -> RotateAmount

buildFromPatternList :: (ExpressionMap, Int) -> Match -> PatternList -> [ExpressionDiff]

buildFromPattern :: (ExpressionMap, Int) -> Match -> Pattern -> ExpressionDiff