HashedExpression.Internal.ToC

Documentation

data MemMap

Constructors

MemMap

Fields

entryMap :: IntMap MemMapEntry

totalDoubles :: Int

type MemMapEntry = (Int, EntryType, Shape)

Mem map (offset, R or C, shape)

type Code = [CodeLine]

makeMemMap :: ExpressionMap -> MemMap

Make a memory map from an expression

makeProblemMemMap :: ExpressionMap -> [Int] -> MemMap

Make a memory map for a problem Make sure varIds are allocated one after another

memOffset :: HasCallStack => MemMap -> Int -> LookupPart -> Int

generateAssignValueCodes :: ValMaps -> MemMap -> ExpressionMap -> Code

Code to assign values to those in val maps

generateEvaluatingCodes :: MemMap -> (ExpressionMap, [Int]) -> Code

Generate evaluation code (usually an expression and its partial derivatives) given an ExpressionMap and indices of nodes to be computed

scoped :: Code -> Code

space :: Int -> Code -> Code

getMemOffsetReal :: HasCallStack => MemMap -> Int -> Int

Get mem offset of real node

data LookupPart

Constructors

LookupR

LookupReC

LookupImC

(<<-) :: String -> String -> CodeLine infixl 1

Helpers to write more readable code generation

i :: String

singleExpressionCProgram :: (DimensionType d, NumType et) => ValMaps -> Expression d et -> Code

Generate a fully working C program that compute the expression and print out the result

localOffset :: [Int] -> [Int] -> Int

Compute the inner offset: e.g localOffset [3, 4, 5] [2, 0, 1] = 2 * (4 * 5) + 0 * (5) + 1