HashedExpression.Internal.ToLLVM

Description

This module has the following functions, LLVMMemMap - is used to allocate memory for all nodes

GenerateEvaluatingCodes - is used to generate LLVM Definition for every operation

mkModule - is used to bind all generated LLVM Definitions to a single LLVM module

Externals - is used to declare all inbuilt functions in LLVM and other helper functions for the above functions.

The two main functions for generating the LLVM IR evaluation functions from an Expression are makeMemoryMap and generateEvaluatingCode.

The makeLLVMMemoryMap function allocates memory for all subexpressions. Expression is a DAG (Directed Acyclic Graph) of subexpressions which is stored in a hash map. Every node contains an operation, references to inputs and its dimensions, and whether it is real or complex. The makeLLVMMemoryMap function gets the list of nodes, calculates its size based on the shape obtained from the Expression map and lays them out linearly. This is implemented with a fold operation. The value accumulated by the fold is a pair of the total size and an IntMap mapping the node id to the offset in the allocated memory.

The generateEvaluatingCode function is used to generate the evaluation function for the given expression. The code is generated one subexpression at a time.

Since the subexpression is a DAG, there is an ordering of the nodes in which inputs occur before subexpressions which use them. Say for instance, in the expression 2x + 2y, the subexpressions can be calculated in the order 2x,2y,2x + 2y, but 2x + 2y could not calculated before 2*y, because it needs it as an input. This type of order is called a topological order, and is implemented in topologicalSortManyRoots. Many roots means that it can be used to evaluate multiple expressions with shared subexpressions.

Code generation is implemented by taking the sorting keys and mapping an evaluation function over the list of keys. The evaluation function matches the operation to LLVM IR instructions.

NOTE: The test suite for this module is available at testTestToLLVM directory. Please go through the README.md file in this directory for running the test cases.

Documentation

elemType :: Type

all elements used here are of double type

type LLVMMemMapEntry = (Int, EntryType, Shape)

Every node entry in the memory has an Index, Type of entry which can be either Real or Complex and its dimension

type Code

Arguments

= [Named Instruction]
List of LLVM instructions forming final code

LLVM Code is a list of Named instructions like Add, FAdd, Mul, Call etc.

data LLVMMemMap

LLVMMemMap has the every node with an index

Constructors

LLVMMemMap

Fields

entryMap :: IntMap LLVMMemMapEntry

every node entry with its index.

totalDoubles :: Int

Total number of entries

makeLLVMMemMap

Arguments

:: ExpressionMap
Expression with its dimension/shape

-> LLVMMemMap
node entry in memory

makeLLVMMemMap takes any expression map and converts it to corresponding LLVM memory map

mmUpdate

Arguments

:: ExpressionMap
Expression mapped with its dimension

-> (IntMap LLVMMemMapEntry, Int)
current memory and its size

-> Int
node id

-> (IntMap LLVMMemMapEntry, Int)
updated memory and updated size

mmUpdate is a helper function for |makeLLVMMemMap|

generateEvaluatingCodes

Arguments

:: String
Name of the function to be generated in LLVM

-> LLVMMemMap
mapped node from memory

-> (ExpressionMap, [Int])
expression map and its indices

-> Definition
LLVM definition for the given expression

Generate evaluation code (usually an expression and its partial derivatives) given an ExpressionMap and indices of nodes to be computed

funcType :: Type

defining type of functions

nameDef

Arguments

:: String
Name of the external function in terms of string

-> Name
Name of the external function in LLVM Name Type

Function to generate LLVM IR to declare the signature of an external function. | nameDef -> String -> AST.Name

mkModule

Arguments

:: String
name for the LLVM module to be generated

-> Expression d et
It is the node id for the final expression. `d` is the expressionID and `et` is the expressionMap

-> Module
LLVM module that has all generated LLVM Definitions

mkModule function for combining all generated LLVM Definitions to a single LLVM module | mkModule :: String -> Expression d et -> AST.Module

externals

Arguments

:: [Definition]
List of LLVM Definitions

Declaration of all functions used from the built-in math library.