HashedExpression.Operation

Documentation

fromDouble :: forall d. ToShape d => Double -> Expression d R

huber :: forall d. DimensionType d => Double -> Expression d R -> Expression d R

Huber loss: https://en.wikipedia.org/wiki/Huber_loss

norm2 :: DimensionType d => Expression d R -> Expression Scalar R

Norm 2

norm1 :: DimensionType d => Expression d R -> Expression Scalar R

class Norm2SquareOp a b | a -> b where

Norm 2 square

Methods

norm2square :: a -> b

huberNorm :: DimensionType d => Double -> Expression d R -> Expression Scalar R

sumElements :: forall d. DimensionType d => Expression d R -> Expression Scalar R

Discrete fourier transform

| Sum across

valueFromNat :: forall n. KnownNat n => Int

variable :: String -> Expression Scalar R

Create primitive expressions

variable1D :: forall n. KnownNat n => String -> Expression n R

Create primitive expressions using Nat kind

variable2D :: forall m n. (KnownNat m, KnownNat n) => String -> Expression '(m, n) R

variable3D :: forall m n p. (KnownNat m, KnownNat n, KnownNat p) => String -> Expression '(m, n, p) R

constant :: Double -> Expression Scalar R

constant1D :: forall n. KnownNat n => Double -> Expression n R

constant2D :: forall m n. (KnownNat m, KnownNat n) => Double -> Expression '(m, n) R

constant3D :: forall m n p. (KnownNat m, KnownNat n, KnownNat p) => Double -> Expression '(m, n, p) R

Orphan instances

(ElementType et, KnownNat n) => RotateOp Int (Expression n et)

Instance details

Methods

rotate :: Int -> Expression n et -> Expression n et

(ElementType et, KnownNat m, KnownNat n) => RotateOp (Int, Int) (Expression ((,) m n) et)

Instance details

Methods

rotate :: (Int, Int) -> Expression (m, n) et -> Expression (m, n) et

(ElementType et, KnownNat m, KnownNat n, KnownNat p) => RotateOp (Int, Int, Int) (Expression ((,,) m n p) et)

Instance details

Methods

rotate :: (Int, Int, Int) -> Expression (m, n, p) et -> Expression (m, n, p) et

ToShape d => Floating (Expression d R)

Instance details

Methods

pi :: Expression d R

exp :: Expression d R -> Expression d R

log :: Expression d R -> Expression d R

sqrt :: Expression d R -> Expression d R

(**) :: Expression d R -> Expression d R -> Expression d R

logBase :: Expression d R -> Expression d R -> Expression d R

sin :: Expression d R -> Expression d R

cos :: Expression d R -> Expression d R

tan :: Expression d R -> Expression d R

asin :: Expression d R -> Expression d R

acos :: Expression d R -> Expression d R

atan :: Expression d R -> Expression d R

sinh :: Expression d R -> Expression d R

cosh :: Expression d R -> Expression d R

tanh :: Expression d R -> Expression d R

asinh :: Expression d R -> Expression d R

acosh :: Expression d R -> Expression d R

atanh :: Expression d R -> Expression d R

log1p :: Expression d R -> Expression d R

expm1 :: Expression d R -> Expression d R

log1pexp :: Expression d R -> Expression d R

log1mexp :: Expression d R -> Expression d R

ToShape d => Fractional (Expression d C)

Instance details

Methods

(/) :: Expression d C -> Expression d C -> Expression d C

recip :: Expression d C -> Expression d C

fromRational :: Rational -> Expression d C

ToShape d => Fractional (Expression d R)

Instance details

Methods

(/) :: Expression d R -> Expression d R -> Expression d R

recip :: Expression d R -> Expression d R

fromRational :: Rational -> Expression d R

ToShape d => Num (Expression d Covector)

Instance details

Methods

(+) :: Expression d Covector -> Expression d Covector -> Expression d Covector

(-) :: Expression d Covector -> Expression d Covector -> Expression d Covector

(*) :: Expression d Covector -> Expression d Covector -> Expression d Covector

negate :: Expression d Covector -> Expression d Covector

abs :: Expression d Covector -> Expression d Covector

signum :: Expression d Covector -> Expression d Covector

fromInteger :: Integer -> Expression d Covector

ToShape d => Num (Expression d C)

Instance details

Methods

(+) :: Expression d C -> Expression d C -> Expression d C

(-) :: Expression d C -> Expression d C -> Expression d C

(*) :: Expression d C -> Expression d C -> Expression d C

negate :: Expression d C -> Expression d C

abs :: Expression d C -> Expression d C

signum :: Expression d C -> Expression d C

fromInteger :: Integer -> Expression d C

ToShape d => Num (Expression d R)

Instance details

Methods

(+) :: Expression d R -> Expression d R -> Expression d R

(-) :: Expression d R -> Expression d R -> Expression d R

(*) :: Expression d R -> Expression d R -> Expression d R

negate :: Expression d R -> Expression d R

abs :: Expression d R -> Expression d R

signum :: Expression d R -> Expression d R

fromInteger :: Integer -> Expression d R

(DimensionType d, NumType et) => PowerOp (Expression d et) Int

Instance details

Methods

(^) :: Expression d et -> Int -> Expression d et

DimensionType d => FTOp (Expression d R) (Expression d C)

Instance details

Methods

ft :: Expression d R -> Expression d C

DimensionType d => FTOp (Expression d C) (Expression d C)

Instance details

Methods

ft :: Expression d C -> Expression d C

(DimensionType d, ElementType et) => PiecewiseOp (Expression d R) (Expression d et)
Piecewise, with a condition expression and branch expressions This is element corresponding, so condition and all branches should have the same dimension and shape

Instance details

Methods

piecewise :: [Double] -> Expression d R -> [Expression d et] -> Expression d et

DimensionType d => ComplexRealOp (Expression d R) (Expression d C)

Instance details

Methods

(+:) :: Expression d R -> Expression d R -> Expression d C

xRe :: Expression d C -> Expression d R

xIm :: Expression d C -> Expression d R

VectorSpace d et s => VectorSpaceOp (Expression Scalar s) (Expression d et)
Scale in vector space

Instance details

Methods

scale :: Expression Scalar s -> Expression d et -> Expression d et

(*.) :: Expression Scalar s -> Expression d et -> Expression d et

InnerProductSpace d s => InnerProductSpaceOp (Expression d s) (Expression d s) (Expression Scalar s)
inner product

Instance details

Methods

(<.>) :: Expression d s -> Expression d s -> Expression Scalar s