HashedExpression.Solver

Documentation

data Problem

Constructors

Problem

Fields

variables :: [Variable]

objectiveId :: Int

expressionMap :: ExpressionMap

memMap :: MemMap

boxConstraints :: [BoxConstraint]

scalarConstraints :: [ScalarConstraint]

data Variable

Constructors

Variable

Fields

varName :: String

nodeId :: Int

partialDerivativeId :: Int

data Constraint

Constructors

NoConstraint

Constraint [ConstraintStatement]

(.>=) :: DimensionType d => Expression d R -> Val -> ConstraintStatement infix 1

(.==) :: DimensionType d => Expression d R -> Val -> ConstraintStatement infix 1

(.<=) :: DimensionType d => Expression d R -> Val -> ConstraintStatement infix 1

between :: DimensionType d => Expression d R -> (Val, Val) -> ConstraintStatement infix 1

data ProblemResult

Constructors

ProblemValid Problem

ProblemInvalid String

NoVariables

constructProblem :: Expression Scalar R -> [String] -> Constraint -> ProblemResult

Construct a Problem from given objective function and constraints

generateProblemCode :: ValMaps -> Problem -> GenResult

data GenResult

Constructors

Invalid String

Success (String -> IO ())

data ConstraintStatement

Constructors

Lower (ExpressionMap, Int) Val

Upper (ExpressionMap, Int) Val

Between (ExpressionMap, Int) (Val, Val)

partialDerivativeMaps :: Expression Scalar Covector -> Map String Int

Return a map from variable name to the corresponding partial derivative node id Partial derivatives in Expression Scalar Covector should be collected before passing to this function

constraintPartialDerivatives :: ScalarConstraint -> [Int]